+++ sites/ademe/CHANGELOG.md
@@ -9,6 +9,10 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Changed
+
+- Migrate to new STORAGES settings
+
 ## [0.20.1] - 2024-04-18
 
 ### Fixed